requirements.txt
tensorflow>=2.19.0
numpy>=1.26.0
pandas>=2.3.0
scikit-learn>=1.7.0
streamlit>=1.28.0
pillow>=10.0.0
matplotlib>=3.7.0
seaborn>=0.12.0 